requirements.txt
pyactiveresource==2.2.2
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
ShopifyAPI==12.6.0
six==1.16.0
